+++ src/main/java/com/guba/spring/speakappbackend/repositories/database/models/Patient.java
@@ -1,40 +1,35 @@
-package com.guba.spring.speakappbackend.database.models;
+package com.guba.spring.speakappbackend.repositories.database.models;
 
 import com.guba.spring.speakappbackend.security.dtos.SignUpDTO;
-import com.guba.spring.speakappbackend.web.schemas.ProfessionalDTO;
+import com.guba.spring.speakappbackend.web.schemas.PatientDTO;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
-import java.util.HashSet;
-import java.util.Set;
 
 @Entity
-@Table(name = "tm_professional")
+@Table(name = "tm_patient")
 @Setter
 @Getter
 @NoArgsConstructor
-public class Professional extends UserAbstract {
+public class Patient extends UserAbstract {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id_professional")
-    private Long idProfessional;
+    @Column(name = "id_patient")
+    private Long idPatient;
 
-    @Column(name = "code")
-    private String code;
+    @ManyToOne
+    @JoinColumn(name="id_professional", nullable=false)
+    private Professional professional;
 
-    @OneToMany(mappedBy="professional")
-    private Set<Patient> patients = new HashSet<>();
-
-    public Professional(ProfessionalDTO p, String password, Role role, LocalDateTime createdAt, LocalDateTime updateAt, Set<Patient> patients) {
+    public Patient(PatientDTO p, String password, Role role, LocalDateTime createdAt, LocalDateTime updateAt, Professional professional) {
         super();
-        this.setIdProfessional(p.getIdProfessional());
+        this.setIdPatient(p.getIdPatient());
         this.setUsername(p.getUsername());
         this.setPassword(password);
-        this.setCode(p.getCode());
         this.setEmail(p.getEmail());
         this.setFirstName(p.getFirstName());
         this.setLastName(p.getLastName());
@@ -44,13 +39,12 @@ public class Professional extends UserAbstract {
         this.setRole(role);
         this.setCreatedAt(createdAt);
         this.setUpdatedAt(updateAt);
-        this.setPatients(patients);
+        this.setProfessional(professional);
     }
 
-    public Professional(SignUpDTO p, String password, String code, Role role, LocalDateTime createdAt, LocalDateTime updateAt) {
+    public Patient(SignUpDTO p, String password, Role role, LocalDateTime createdAt, LocalDateTime updateAt, Professional professional) {
         super();
         this.setUsername(p.getUsername());
-        this.setCode(code);
         this.setPassword(password);
         this.setEmail(p.getEmail());
         this.setFirstName(p.getFirstName());
@@ -58,5 +52,6 @@ public class Professional extends UserAbstract {
         this.setRole(role);
         this.setCreatedAt(createdAt);
         this.setUpdatedAt(updateAt);
+        this.setProfessional(professional);
     }
 }
