+++ src/main/java/com/guba/spring/speakappbackend/repositories/database/models/Professional.java
@@ -1,35 +1,40 @@
-package com.guba.spring.speakappbackend.database.models;
+package com.guba.spring.speakappbackend.repositories.database.models;
 
 import com.guba.spring.speakappbackend.security.dtos.SignUpDTO;
-import com.guba.spring.speakappbackend.web.schemas.PatientDTO;
+import com.guba.spring.speakappbackend.web.schemas.ProfessionalDTO;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
+import java.util.HashSet;
+import java.util.Set;
 
 @Entity
-@Table(name = "tm_patient")
+@Table(name = "tm_professional")
 @Setter
 @Getter
 @NoArgsConstructor
-public class Patient extends UserAbstract {
+public class Professional extends UserAbstract {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id_patient")
-    private Long idPatient;
+    @Column(name = "id_professional")
+    private Long idProfessional;
 
-    @ManyToOne
-    @JoinColumn(name="id_professional", nullable=false)
-    private Professional professional;
+    @Column(name = "code")
+    private String code;
 
-    public Patient(PatientDTO p, String password, Role role, LocalDateTime createdAt, LocalDateTime updateAt, Professional professional) {
+    @OneToMany(mappedBy="professional")
+    private Set<Patient> patients = new HashSet<>();
+
+    public Professional(ProfessionalDTO p, String password, Role role, LocalDateTime createdAt, LocalDateTime updateAt, Set<Patient> patients) {
         super();
-        this.setIdPatient(p.getIdPatient());
+        this.setIdProfessional(p.getIdProfessional());
         this.setUsername(p.getUsername());
         this.setPassword(password);
+        this.setCode(p.getCode());
         this.setEmail(p.getEmail());
         this.setFirstName(p.getFirstName());
         this.setLastName(p.getLastName());
@@ -39,12 +44,13 @@ public class Patient extends UserAbstract {
         this.setRole(role);
         this.setCreatedAt(createdAt);
         this.setUpdatedAt(updateAt);
-        this.setProfessional(professional);
+        this.setPatients(patients);
     }
 
-    public Patient(SignUpDTO p, String password, Role role, LocalDateTime createdAt, LocalDateTime updateAt, Professional professional) {
+    public Professional(SignUpDTO p, String password, String code, Role role, LocalDateTime createdAt, LocalDateTime updateAt) {
         super();
         this.setUsername(p.getUsername());
+        this.setCode(code);
         this.setPassword(password);
         this.setEmail(p.getEmail());
         this.setFirstName(p.getFirstName());
@@ -52,6 +58,5 @@ public class Patient extends UserAbstract {
         this.setRole(role);
         this.setCreatedAt(createdAt);
         this.setUpdatedAt(updateAt);
-        this.setProfessional(professional);
     }
 }
