+++ src/main/java/com/guba/spring/speakappbackend/repositories/database/models/TaskItem.java
@@ -1,4 +1,4 @@
-package com.guba.spring.speakappbackend.database.models;
+package com.guba.spring.speakappbackend.repositories.database.models;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
