+++ src/main/java/com/guba/spring/speakappbackend/repositories/database/repositories/ImageTempRepository.java
@@ -1,6 +1,6 @@
-package com.guba.spring.speakappbackend.database.repositories;
+package com.guba.spring.speakappbackend.repositories.database.repositories;
 
-import com.guba.spring.speakappbackend.database.models.ImageTemp;
+import com.guba.spring.speakappbackend.repositories.database.models.ImageTemp;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
