+++ src/main/java/com/guba/spring/speakappbackend/repositories/database/repositories/PatientRepository.java
@@ -1,6 +1,6 @@
-package com.guba.spring.speakappbackend.database.repositories;
+package com.guba.spring.speakappbackend.repositories.database.repositories;
 
-import com.guba.spring.speakappbackend.database.models.Patient;
+import com.guba.spring.speakappbackend.repositories.database.models.Patient;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
