+++ src/main/java/com/guba/spring/speakappbackend/repositories/database/repositories/RoleRepository.java
@@ -1,6 +1,6 @@
-package com.guba.spring.speakappbackend.database.repositories;
+package com.guba.spring.speakappbackend.repositories.database.repositories;
 
-import com.guba.spring.speakappbackend.database.models.Role;
+import com.guba.spring.speakappbackend.repositories.database.models.Role;
 import com.guba.spring.speakappbackend.enums.RoleEnum;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
