+++ src/main/java/com/guba/spring/speakappbackend/services/ImageService.java
@@ -1,7 +1,7 @@
 package com.guba.spring.speakappbackend.services;
 
-import com.guba.spring.speakappbackend.database.models.Image;
-import com.guba.spring.speakappbackend.database.repositories.ImageRepository;
+import com.guba.spring.speakappbackend.repositories.database.models.Image;
+import com.guba.spring.speakappbackend.repositories.database.repositories.ImageRepository;
 import com.opencsv.bean.CsvBindByName;
 import com.opencsv.bean.CsvToBeanBuilder;
 import lombok.Getter;
