+++ src/main/java/com/guba/spring/speakappbackend/services/ProfessionalService.java
@@ -1,9 +1,9 @@
 package com.guba.spring.speakappbackend.services;
 
-import com.guba.spring.speakappbackend.database.models.Professional;
-import com.guba.spring.speakappbackend.database.models.Role;
-import com.guba.spring.speakappbackend.database.repositories.ProfessionalRepository;
-import com.guba.spring.speakappbackend.database.repositories.RoleRepository;
+import com.guba.spring.speakappbackend.repositories.database.models.Professional;
+import com.guba.spring.speakappbackend.repositories.database.models.Role;
+import com.guba.spring.speakappbackend.repositories.database.repositories.ProfessionalRepository;
+import com.guba.spring.speakappbackend.repositories.database.repositories.RoleRepository;
 import com.guba.spring.speakappbackend.enums.RoleEnum;
 import com.guba.spring.speakappbackend.exceptions.NotFoundElementException;
 import com.guba.spring.speakappbackend.exceptions.NotSavedElementException;
