+++ src/main/java/com/guba/spring/speakappbackend/services/strategies/ResolveConsonantalSyllable.java
@@ -1,6 +1,6 @@
 package com.guba.spring.speakappbackend.services.strategies;
 
-import com.guba.spring.speakappbackend.database.models.TaskItem;
+import com.guba.spring.speakappbackend.repositories.database.models.TaskItem;
 import com.guba.spring.speakappbackend.enums.ResultExercise;
 import com.guba.spring.speakappbackend.web.schemas.ResultExerciseDTO;
 import lombok.RequiredArgsConstructor;
