+++ src/main/java/com/guba/spring/speakappbackend/services/strategies/ResolveStrategy.java
@@ -1,6 +1,6 @@
 package com.guba.spring.speakappbackend.services.strategies;
 
-import com.guba.spring.speakappbackend.database.models.TaskItem;
+import com.guba.spring.speakappbackend.repositories.database.models.TaskItem;
 import com.guba.spring.speakappbackend.web.schemas.ResultExerciseDTO;
 
 public interface ResolveStrategy {
