+++ src/main/java/com/guba/spring/speakappbackend/web/schemas/PhonemeDTO.java
@@ -1,6 +1,6 @@
 package com.guba.spring.speakappbackend.web.schemas;
 
-import com.guba.spring.speakappbackend.database.models.Phoneme;
+import com.guba.spring.speakappbackend.repositories.database.models.Phoneme;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
