+++ src/main/java/com/guba/spring/speakappbackend/web/schemas/ProfessionalDTO.java
@@ -1,6 +1,6 @@
 package com.guba.spring.speakappbackend.web.schemas;
 
-import com.guba.spring.speakappbackend.database.models.Professional;
+import com.guba.spring.speakappbackend.repositories.database.models.Professional;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
